recognize_celebrities/pyproject.toml
[project]
name = "recognize-celebrities"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3-stubs[rekognition]>=1.35.76",
    "boto3>=1.35.76",
    "pillow>=11.0.0",
]
